feat: migrate entry points to TypeScript
bot.ts replaces module augmentation with UbotClient extends Client.
deploy-commands.ts adds exit guards for env vars and explicit command
array type.
